requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
streamlit==1.39.0
sqlalchemy==2.0.35
pydantic==2.9.2
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
textblob==0.18.0.post0
plotly==5.24.1
pydeck==0.9.1
requests==2.32.3
python-dateutil==2.9.0.post0
groq>=0.33.0
httpx==0.27.2
python-dotenv==1.0.1
pinecone>=7.3.0
sentence-transformers==2.2.2

